MBS-11453: Change entity0_cardinality, entity1_cardinality to SMALLINT
Values are 0-1 and will certainly never go over 30k.

+++ admin/sql/CreateTables.sql
@@ -2264,8 +2264,8 @@ CREATE TABLE link_type ( -- replicate
     last_updated        TIMESTAMP WITH TIME ZONE DEFAULT NOW(),
     is_deprecated       BOOLEAN NOT NULL DEFAULT false,
     has_dates           BOOLEAN NOT NULL DEFAULT true,
-    entity0_cardinality INTEGER NOT NULL DEFAULT 0,
-    entity1_cardinality INTEGER NOT NULL DEFAULT 0
+    entity0_cardinality SMALLINT NOT NULL DEFAULT 0,
+    entity1_cardinality SMALLINT NOT NULL DEFAULT 0
 );
 
 CREATE TABLE link_type_attribute_type ( -- replicate
